requirements.txt
colorama==0.4.6
Flask==3.1.1
flask_migrate==4.1.0
flask_restx==1.3.0
flask_sqlalchemy==3.1.1
openpyxl==3.1.5
pandas==2.2.3
psutil==7.0.0
PyMySQL==1.1.1
pytest==8.3.5
pytest-cov==6.1.1
python-dotenv==1.1.0
Requests==2.32.3
SQLAlchemy==2.0.41
sqlalchemy_utils==0.41.2
